pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-rag-agent"
version = "0.1.0"
description = "An adaptive RAG agent that autonomously decides retrieval strategies. Supports URL crawling and local docs."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["rag", "agent", "langchain", "langgraph", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.26",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
rerank = []  # Reranker uses sentence-transformers (already in dependencies)
eval = ["ragas>=0.2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
all = ["smart-rag-agent[rerank,eval,dev]"]

[project.scripts]
smart-rag = "smart_rag.cli:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/smart-rag-agent"
Repository = "https://github.com/YOUR_USERNAME/smart-rag-agent"
Issues = "https://github.com/YOUR_USERNAME/smart-rag-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
